server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cbor2>=5.7.1",
    "dj-database-url>=3.0.1",
    "django>=6.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "fido2>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "ruff>=0.14.8",
]
